marka ya göre listeleme getirildi

+++ lib/pages/page_details.dart
@@ -1,7 +1,9 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:carpricemobile/design_config/color.dart';
+import 'package:carpricemobile/pages/profile_page.dart';
 import 'package:carpricemobile/services/firestore_service.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 
@@ -18,6 +20,7 @@ class ProductDetails extends StatefulWidget {
   late String price;
   late bool isLiked;
   late int likeCount;
+  late String updateDate;
 
   ProductDetails(
       {super.key,
@@ -30,7 +33,8 @@ class ProductDetails extends StatefulWidget {
       required this.description,
       required this.price,
       required this.isLiked,
-      required this.likeCount});
+      required this.likeCount,
+      required this.updateDate});
 
   @override
   State<ProductDetails> createState() => _ProductDetailsState();
@@ -40,7 +44,9 @@ class _ProductDetailsState extends State<ProductDetails> {
   @override
   void initState() {
     widget.isLiked;
-    print("init State isLiked =>  ${widget.isLiked}");
+    if (kDebugMode) {
+      print("init State isLiked =>  ${widget.isLiked}");
+    }
     super.initState();
   }
 
@@ -50,11 +56,25 @@ class _ProductDetailsState extends State<ProductDetails> {
 
 
   Future<void> addToFavorite() async {
-      widget.isLiked = true;
+    if (user == null) {
+      beLogin();
 
-
-    addFavorite.doc("${user?.uid}${widget.aracID}").set({'AracID': widget.aracID, 'FavID': addFavorite.id + user!.uid, 'UserEmail': user!.email, 'UserID': user!.uid, 'AracResimUrl': widget.imagePath, 'AracFiyat': widget.price, 'AracOzellikleri':widget.description,
-      'Marka': widget.name, 'Model': widget.model, 'KasaTipi': widget.caseType, 'VasitaTipi': widget.vehicleType, 'isLiked': widget.isLiked,
+    } else {
+    widget.isLiked = true;
+    addFavorite.doc("${user?.uid}${widget.aracID}").set({
+      'AracID': widget.aracID,
+      'FavID': addFavorite.id + user!.uid,
+      'UserEmail': user!.email,
+      'UserID': user!.uid,
+      'AracResimUrl': widget.imagePath,
+      'AracFiyat': widget.price,
+      'AracOzellikleri': widget.description,
+      'Marka': widget.name,
+      'Model': widget.model,
+      'KasaTipi': widget.caseType,
+      'VasitaTipi': widget.vehicleType,
+      'isLiked': widget.isLiked,
+      'eklenmeTarihi': widget.updateDate,
     }).then((value) {
       setState(() {
         widget.isLiked = true;
@@ -67,11 +87,13 @@ class _ProductDetailsState extends State<ProductDetails> {
         addToFav();
       });
     });
-
+   }
   }
 
   Future deleteToFavorite() async {
-
+    if(user == null){
+      beLogin();
+    }else{
       addFavorite.doc("${user?.uid}${widget.aracID}").delete().then((value) {
         if(widget.likeCount== 0 || widget.likeCount <= 0){
           widget.isLiked = false;
@@ -88,8 +110,14 @@ class _ProductDetailsState extends State<ProductDetails> {
                 merge: true,
               ));
         });
-        print("foviri silindi");
-      });
+        if (kDebugMode) {
+          print("favori silindi");
+            }
+          }
+      );
+    }
+
+
 
 
 
@@ -123,12 +151,22 @@ class _ProductDetailsState extends State<ProductDetails> {
                   crossAxisAlignment: CrossAxisAlignment.center,
                   children: [
                     Padding(
+                      padding: const EdgeInsets.all(5.0),
+                      child: Row(mainAxisAlignment: MainAxisAlignment.end,
+                        children: [
+                          RichText(text: TextSpan(spellOut: true,text: "  Güncelleme Tarihi  ${widget.updateDate.substring(0,10)}",style: Theme.of(context).textTheme.labelMedium)),
+                        ],
+                      ),
+                    ),
+                    Padding(
                       padding: const EdgeInsets.all(8.0),
                       child: RichText(
+                        overflow: TextOverflow.ellipsis,
                         text: TextSpan(
                             text: "${widget.name} ",
                             children: [
-                              TextSpan(text: widget.model, style: Theme.of(context).textTheme.titleMedium)
+                              TextSpan(text: widget.model, style: Theme.of(context).textTheme.titleMedium),
+
                             ],
                             style: Theme.of(context).textTheme.titleLarge),
                       ),
@@ -173,8 +211,10 @@ class _ProductDetailsState extends State<ProductDetails> {
                                 alignment: AlignmentDirectional.topStart,
                                 backgroundColor: MyColors().scaffoldBackground,
                                 label: Text("${widget.likeCount}",style: Theme.of(context).textTheme.labelSmall,),
-                                child:widget.likeCount ==1 || widget.likeCount >=1 ? Icon(Icons.favorite, size: 40, color: Colors.red.shade600)
+                                child:widget.likeCount ==1 || widget.likeCount >=1 || widget.isLiked ?
+                                Icon(Icons.favorite, size: 40, color: Colors.red.shade600)
                                     :
+
                                 Icon(Icons.mood_bad, size: 40, color: Colors.blue.shade800,)
                               ),
                           ),
@@ -183,18 +223,33 @@ class _ProductDetailsState extends State<ProductDetails> {
                     ),
                     divider(),
                     Row(
-                      mainAxisAlignment: MainAxisAlignment.end,
+                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                       children: [
-                        // IconButton(onPressed: (){
-                        //   Navigator.pushNamed(context, '/favoriler');
-                        // }, icon: const Icon(CupertinoIcons.heart)),
+                        TextButton(onPressed: (){
+                          Navigator.pushNamed(context, '/favoriler');
+                        }, child: Text("Favoriler",style: Theme.of(context).textTheme.labelLarge,)),
                         TextButton(
                             onPressed: () {
                               Navigator.pushNamed(context, '/home',);
                             },
-                            child:  Text("Tüm Araçları Gör",style: Theme.of(context).textTheme.titleMedium)),
+                            child:  Text("Tüm Araçlar",style: Theme.of(context).textTheme.labelLarge)),
+                        TextButton(
+                            onPressed: () {
+                              Navigator.push(context, MaterialPageRoute(builder: (context)=>const ProfilPage()));
+                            },
+                            child:  Text("Profil",style: Theme.of(context).textTheme.labelLarge)),
                       ],
                     ),
+                    SingleChildScrollView(scrollDirection: Axis.horizontal,
+                      child: Row(
+                        children: [
+                          Padding(
+                            padding: const EdgeInsets.all(5.0),
+                            child: RichText(text: TextSpan(text: "Favorilemek için 2 kez dokunun  geri almak içinse kalp sembolüne 1 kez dokunun",style: Theme.of(context).textTheme.labelSmall,spellOut:true,)),
+                          ),
+                        ],
+                      ),
+                    )
                   ],
                 ),
               ),
@@ -219,6 +274,14 @@ class _ProductDetailsState extends State<ProductDetails> {
         fontSize: 15,
         webPosition: "center");
     const Duration(milliseconds: 1000);
+  }
+  void beLogin() async {
+    await Fluttertoast.showToast(
+        msg: "Lütfen üye girişi yapın",
+        // webBgColor: "linear-gradient(to right, #2DE5E5, #2DE5E5)",
+        fontSize: 15,
+        webPosition: "center");
+    const Duration(seconds: 3);
   }
 
   void deleteToFav() async {
